Add simple customer api with swagger

See merge request adeev/web-shop!2

+++ nfo-shop/src/main/resources/db/changelog/v-1.0/10-create-products-table.sql
@@ -2,7 +2,7 @@ create table products
 (
     id          bigserial not null,
     name        varchar(100),
-    price       int,
+    price       double precision,
     description text,
     primary key (id)
 );
